+++ webworker/import-eval-path.ts
@@ -16,7 +16,7 @@ import {
   getPackageJsonEntrypoint,
   isTypeScriptEntrypoint,
   isDistDirEmpty,
-} from "./package-validation"
+} from "./index"
 
 const debug = Debug("tsci:eval:import-eval-path")
