+++ webworker/index.ts
@@ -1,1 +1,11 @@
+/**
+ * Package validation utilities
+ * Each function is in its own file for better organization and modularity
+ */
 export * from "./entrypoint"
+export { extractBasePackageName } from "./extractBasePackageName"
+export { isPackageDeclaredInPackageJson } from "./isPackageDeclaredInPackageJson"
+export { getNodeModuleDirectory } from "./getNodeModuleDirectory"
+export { getPackageJsonEntrypoint } from "./getPackageJsonEntrypoint"
+export { isTypeScriptEntrypoint } from "./isTypeScriptEntrypoint"
+export { isDistDirEmpty } from "./isDistDirEmpty"
